Test: 카테고리와 상품 테스트 통합
